Rename main app file
